FinalYearProject/requirements.txt
flask
requests
pillow
rembg
boto3
